Split of BuildOnWorkerITest into two Files, final modifications
